Create README.md
Churn rate is the process of measuring the rate at which customers stop using a certain product. In the telecommunication industry churn analysis helps these companies understand their customers better and know how they should improve the services they are offering in order to reduce churn. The lower the churn rate the better